ordering/src/routes/show.ts
import {
  NotAuthorizedError,
  NotFoundError,
  requireAuth,
} from "@microservices-with-react-and-node-sharmarajdaksh/common";
import express from "express";
import { Order } from "../models/order";

const router = express.Router();

router.get(
  "/api/orders/:orderId",
  requireAuth,
  async (req: express.Request, res: express.Response) => {
    const { orderId } = req.params;

    const order = await Order.findById(orderId).populate("ticket");

    if (!order) {
      throw new NotFoundError();
    }

    if (order.userId !== req.user!.id) {
      throw new NotAuthorizedError();
    }

    return res.send(order);
  }
);

export { router as showOrderRouter };
